requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
httpx>=0.27.0
biopython>=1.83
scikit-learn>=1.4.0
numpy>=1.26.0
